requirements.txt
Django>=4.2,<5.1
djangorestframework
djangorestframework-simplejwt
mysqlclient
django_redis
python-dotenv